Second attempt at fixing erro with absolute values including of unsigned numbers...

+++ src/rseg/rseg-diff.cpp
@@ -288,8 +288,8 @@ output_boundaries(const vector<double> &tmp_read_bins,
     {
       static const size_t sz = tmp_boundary_scores.size();
 
-      if (std::abs(i - change_points[j]) >
-          std::abs(i - change_points[j + 1]))
+      if (std::abs(i - static_cast<long int>(change_points[j])) >
+          std::abs(i - static_cast<long int>(change_points[j + 1])))
         ++j;
 
       const size_t trn = change_points[j];
